correct cart_item price from intiger to decimalfield

+++ shop/models/cart_item.py
@@ -14,7 +14,8 @@ class CartItem(models.Model):
     product = models.ForeignKey(Product, related_name='+', on_delete=models.CASCADE)
     quantity = models.IntegerField(default=1)
 
-    price = models.IntegerField(default=0)
+    # price = item_price * quantity
+    price = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
 
     comment =  models.CharField(max_length=255, blank=True)
